sarkin-clip/requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astor==0.8.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
babel==2.17.0
bce-python-sdk==0.9.55
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.2
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
distro==1.9.0
easyocr==1.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.123.10
ffmpy==1.0.0
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jiwer==3.0.4
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mlx==0.29.3
mlx-lm==0.28.4
mlx-metal==0.29.3
mlx-vlm==0.1.15
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.2.1
ninja==1.13.0
numpy==1.26.4
open_clip_torch==3.2.0
openai==1.58.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.4
packaging==25.0
paddlepaddle==3.0.0
pandas==2.3.3
pdf2docx==0.5.8
pillow==10.4.0
portalocker==3.2.0
premailer==3.10.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pytest==8.3.4
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
RapidFuzz==3.13.0
rarfile==4.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruff==0.14.8
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
sympy==1.14.0
termcolor==3.1.0
tifffile==2024.8.30
timm==1.0.22
tokenizers==0.22.1
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
visualdl==2.5.3
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
